refactor: merge `usePresentation()` hook into `SlidesContext`
