+++ src/components/ui/slide/index.tsx
@@ -7,7 +7,6 @@ import { Carousel } from "@qwik-ui/headless";
 import transitions from "./slides-transitions.css?inline";
 import { useDoubleClick } from "~/hooks/use-double-click";
 import type { SlideEntity } from "~/types/SlideEntity";
-import { usePresentation } from "~/hooks/use-presentation";
 
 export const Slide = component$<{ slide: SlideEntity }>((props) => {
   useStyles$(transitions);
@@ -57,7 +56,6 @@ export const Slide = component$<{ slide: SlideEntity }>((props) => {
   `);
 
   const slides = useContext(SlidesContextId);
-  const presentation = usePresentation()
 
   const handleDoubleClick$ = useDoubleClick(
     $(() => {
@@ -78,7 +76,7 @@ export const Slide = component$<{ slide: SlideEntity }>((props) => {
         <Carousel.Slide class="slide-controls">
           <aside role="toolbar" aria-label="Slide controls">
             <Button
-              onClick$={() => presentation.removeSlide(props.slide.id)}
+              onClick$={() => slides.remove(props.slide.id)}
               class="red size-lg"
               tabIndex={-1}
               icon="delete"
