+++ src/contexts/slides-context.tsx
@@ -1,3 +1,4 @@
+import type { QRL } from "@builder.io/qwik";
 import {
   $,
   component$,
@@ -17,6 +18,8 @@ interface SlidesStore {
   _active: SlideEntity | null;
   active: SlideEntity | null;
   array: SlideEntity[];
+  remove: QRL<(id: number) => Promise<void>>;
+  add: QRL<(files: File[]) => Promise<void>>;
 }
 
 export const SlidesContextId = createContextId<SlidesStore>("slides");
@@ -32,24 +35,38 @@ export const SlidesContextProvider = component$(() => {
       this._active = value;
     },
     array: [],
+    remove: $(async () => {}),
+    add: $(async () => {}),
   }));
 
-  useOnWindow('load', $(async () => {
+  useOnWindow(
+    "load",
+    $(async () => {
+      store.add = $(async (files: File[]) => {
+        for (const file of files) {
+          const id = Date.now();
+          await IndexedDBService.saveSlideFile(id, file);
 
-    IndexedDBService.getAllSlides().then((slides) => {
-      slides.forEach((slide) => {
-        const s = {
-          id: slide.id,
-          fileName: slide.file.name,
-          preview: URL.createObjectURL(slide.file),
+          const slide = {
+            id,
+            fileName: file.name,
+            preview: URL.createObjectURL(file),
+          };
+
+          store.array.push(slide);
         }
-        store.array.push(s);
       });
+
+      store.remove = $(async (id: number) => {
+        const index = store.array.findIndex((slide) => slide.id === id);
+        if (index === -1) throw new Error(`Slide with id ${id} not found`);
+        store.array.splice(index, 1);
+        await IndexedDBService.removeSlide(id);
+      });
+
+      store.add(await IndexedDBService.getAllFiles());
     })
-
-  }))
-
-  
+  );
 
   // TODO: Should be merged into one hook,
   //       like useLocalStorage$().
